+++ src/elements/conditions/contacts/LastActivityConditionRule.php
@@ -1,0 +1,54 @@
+<?php
+/**
+ * @copyright Copyright (c) PutYourLightsOn
+ */
+
+namespace putyourlightson\campaign\elements\conditions\contacts;
+
+use Craft;
+use craft\base\conditions\BaseDateRangeConditionRule;
+use craft\base\ElementInterface;
+use craft\elements\conditions\ElementConditionRuleInterface;
+use craft\elements\db\ElementQueryInterface;
+use putyourlightson\campaign\elements\ContactElement;
+use putyourlightson\campaign\elements\db\ContactElementQuery;
+
+/**
+ * @since 2.0.0
+ */
+class LastActivityConditionRule extends BaseDateRangeConditionRule implements ElementConditionRuleInterface
+{
+    /**
+     * @inheritdoc
+     */
+    public function getLabel(): string
+    {
+        return Craft::t('campaign', 'Last Activity');
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getExclusiveQueryParams(): array
+    {
+        return ['lastActivity'];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function modifyQuery(ElementQueryInterface $query): void
+    {
+        /** @var ContactElementQuery $query */
+        $query->lastActivity($this->queryParamValue());
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function matchElement(ElementInterface $element): bool
+    {
+        /** @var ContactElement $element */
+        return $this->matchValue($element->lastActivity);
+    }
+}